requirements.txt
appdirs
inflect
mysqlclient
pandas
ruamel.yaml
SQLAlchemy
sqlalchemy-mixins
SQLAlchemy-Utils
